refactor: create @pnpm/parse-cli-args
