+++ packages/find-workspace-dir/README.md
@@ -1,0 +1,15 @@
+# @pnpm/find-workspace-dir
+
+> Finds the root of a pnpm workspace
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/find-workspace-dir.svg)](https://www.npmjs.com/package/@pnpm/find-workspace-dir)
+
+## Installation
+
+```sh
+<pnpm|npm|yarn> add @pnpm/find-workspace-dir
+```
+
+## License
+
+MIT © [Zoltan Kochan](https://www.kochan.io/)
