+++ packages/find-workspace-dir/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@pnpm/find-workspace-dir",
+  "version": "0.0.0",
+  "description": "Finds the root of a pnpm workspace",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
+    "tsc": "rimraf lib && tsc",
+    "test": "pnpm run tsc && pnpm run lint",
+    "prepublishOnly": "pnpm run tsc"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/find-workspace-dir",
+  "keywords": [
+    "pnpm"
+  ],
+  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://pnpm.js.org",
+  "devDependencies": {
+    "rimraf": "3.0.0"
+  },
+  "dependencies": {
+    "find-up": "4.1.0"
+  }
+}
