+++ packages/find-workspace-dir/src/index.ts
