+++ packages/parse-cli-args/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@pnpm/parse-cli-args",
+  "version": "0.0.0",
+  "description": "Parses the CLI args passed to pnpm",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
+    "tsc": "rimraf lib && tsc",
+    "test": "pnpm run tsc && pnpm run lint",
+    "prepublishOnly": "pnpm run tsc"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/parse-cli-args",
+  "keywords": [
+    "pnpm"
+  ],
+  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://pnpm.js.org",
+  "devDependencies": {
+    "@types/nopt": "3.0.29",
+    "rimraf": "3.0.0"
+  },
+  "dependencies": {
+    "@pnpm/find-workspace-dir": "workspace:0.0.0",
+    "chalk": "3.0.0",
+    "nopt": "4.0.1"
+  }
+}
