+++ packages/pnpm/package.json
@@ -28,7 +28,9 @@
     "@pnpm/default-reporter": "workspace:5.0.3",
     "@pnpm/error": "workspace:1.0.0",
     "@pnpm/file-reporter": "0.1.0",
+    "@pnpm/find-workspace-dir": "workspace:0.0.0",
     "@pnpm/logger": "3.1.0",
+    "@pnpm/parse-cli-args": "workspace:0.0.0",
     "@pnpm/plugin-commands-audit": "workspace:0.0.0",
     "@pnpm/plugin-commands-import": "workspace:0.0.0",
     "@pnpm/plugin-commands-installation": "workspace:0.0.0",
